requirements.txt
# Core LangChain packages
langchain>=0.2.5,<0.3.0
langchain-community>=0.2.5,<0.3.0
langchain-openai>=0.1.7,<0.2.0

# OpenAI
openai>=1.35.0,<2.0.0

# Vector Database
chromadb>=0.4.24,<0.5.0
duckdb>=0.9.0

# Web scraping
beautifulsoup4>=4.12.0
requests>=2.31.0

# Web framework
streamlit>=1.28.0

# Utilities
schedule>=1.2.0
tiktoken>=0.5.0
python-dotenv>=1.0.0

# Free Embeddings Alternative (Optional - only needed if USE_FREE_EMBEDDINGS=true)
sentence-transformers>=2.2.0
torch>=2.0.0

# Free Local LLM Support (Optional - only needed if USE_LOCAL_LLM=true)
# For Ollama: Install Ollama from https://ollama.ai (separate installation)
# For HuggingFace: Install transformers and accelerate
transformers>=4.30.0
accelerate>=0.20.0

# PDF extraction
pypdf>=4.2.0

# Python 3.11.14 compatible
